Fixed health models

+++ src/main/java/com/pjeskiem_i_mieczem/StatBar.java
@@ -12,13 +12,13 @@ import javafx.scene.layout.StackPane;
 public class StatBar extends HBox {
     StackPane bar = new StackPane();
     private int width;
-    float maxValue;
+    double maxValue;
     private Label textLabel = new Label();
     private boolean showText;
 //    Bar without text
 
 //    Bar with progress text
-    public StatBar(String colour, Boolean showText, int width, int height, float maxValue, float currentValue){
+    public StatBar(String colour, Boolean showText, int width, int height, double maxValue, double currentValue){
         this.width = width;
         this.showText = showText;
         bar.setMinHeight(height);
@@ -31,7 +31,7 @@ public class StatBar extends HBox {
         this.setMinWidth(width);
         this.getChildren().add(bar);
     }
-    public void updateBar(float currentValue){
+    public void updateBar(double currentValue){
         bar.setPrefWidth(Math.min(1, currentValue/maxValue)*width);
         if (showText)
             textLabel.setText(Math.round(currentValue) + "/"+ Math.round(maxValue));
